requirements.txt
arize-phoenix>=8.0.0
arize-phoenix-evals>=2.10.0
openai>=1.30.0
openinference-instrumentation-openai>=0.1.0
pandas>=2.2.0
numpy>=1.26.0
datasets>=2.19.0
pyyaml>=6.0.1
scipy>=1.13.0
matplotlib>=3.9.0
seaborn>=0.13.0
scikit-learn>=1.5.0
jupyter>=1.0.0
nbformat>=5.10.0
tqdm>=4.66.0
